feat(chart): @trendcraft/chart/replay subpath with createLiveSimulator
Lifts the live-simulator helper that was duplicated in
examples/indicator-showcase and examples/strategy-studio into a
canonical chart subpath. Drives a `createLiveCandle` instance from a
static candle array on a timer, splitting each pending candle into N
intra-candle ticks before the final candleComplete — useful for any
chart host that wants to demo `connectIndicators({ live })` without
a real market feed.

API:
- `createLiveSimulator(opts)` returns a `SimulatorHandle` with
  `play()`, `pause()`, `stepOnce({ wholeBar? })`, `setIntervalMs`,
  `reset()`, `onChange(cb)`, `dispose()` plus the `live` LiveSource
  that plugs straight into chart's connect APIs.
- Eager completion: both autoplay and stepOnce flip `state` to
  `complete` the moment the last queued candle is emitted, instead
  of lagging one timer interval behind.

Build / publish wiring:
- New entry in `vite.config.ts`, `package.json#exports`,
  `package.json#size-limit` (~3 kB brotli, observed 807 B).
- `scripts/verify-publish.mjs` checks the new subpath exports
  `createLiveSimulator`.

Examples:
- `indicator-showcase/src/live-simulator.ts` becomes a thin re-export
  shim. Vite alias and TS path map updated for source-mode dev.
- Strategy Studio's local copy will be replaced with the same shim
  on the next epic→main merge (Studio code lives on epic, not yet
  on main; aliases there will need the same `@trendcraft/chart/replay`
  entry).

Runtime requirement: `@trendcraft/chart/replay` imports
`createLiveCandle` from `trendcraft`, so consumers using this subpath
must install `trendcraft` alongside `@trendcraft/chart`. The chart's
main entry stays standalone; only `replay` (and existing `presets`)
need the optional peer.

3 rounds of Codex review; final round flagged no defects.
